pyproject.toml
[project]
name = "ollamademo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "ollama>=0.4.8",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
]
